+++ rust/main/chains/hyperlane-sealevel/src/mailbox.rs
@@ -501,6 +501,11 @@ impl Mailbox for SealevelMailbox {
         // If we're using Jito, we need to send a tip to the Jito fee account.
         // Otherwise, we need to set the compute unit price.
         if self.use_jito() {
+            let tip: u64 = std::env::var("JITO_TIP_LAMPORTS")
+                .ok()
+                .and_then(|s| s.parse::<u64>().ok())
+                .unwrap_or(PROCESS_DESIRED_PRIORITIZATION_FEE_LAMPORTS_PER_TX);
+
             // The tip is a standalone transfer to a Jito fee account.
             // See https://github.com/jito-labs/mev-protos/blob/master/json_rpc/http.md#sendbundle.
             instructions.push(solana_sdk::system_instruction::transfer(
@@ -508,7 +513,7 @@ impl Mailbox for SealevelMailbox {
                 // A random Jito fee account, taken from the getFeeAccount RPC response:
                 // https://github.com/jito-labs/mev-protos/blob/master/json_rpc/http.md#gettipaccounts
                 &solana_sdk::pubkey!("DfXygSm4jCyNCybVYYK6DwvWqjKee8pbDmJGcLWNDXjh"),
-                PROCESS_DESIRED_PRIORITIZATION_FEE_LAMPORTS_PER_TX,
+                tip,
             ));
         }
         // "processed" level commitment does not guarantee finality.
@@ -648,10 +653,15 @@ pub struct SealevelMailboxIndexer {
     program_id: Pubkey,
     dispatch_message_log_meta_composer: LogMetaComposer,
     delivery_message_log_meta_composer: LogMetaComposer,
+    advanced_log_meta: bool,
 }
 
 impl SealevelMailboxIndexer {
-    pub fn new(conf: &ConnectionConf, locator: ContractLocator) -> ChainResult<Self> {
+    pub fn new(
+        conf: &ConnectionConf,
+        locator: ContractLocator,
+        advanced_log_meta: bool,
+    ) -> ChainResult<Self> {
         let program_id = Pubkey::from(<[u8; 32]>::from(locator.address));
         let mailbox = SealevelMailbox::new(conf, locator, None)?;
 
@@ -672,6 +682,7 @@ impl SealevelMailboxIndexer {
             mailbox,
             dispatch_message_log_meta_composer,
             delivery_message_log_meta_composer,
+            advanced_log_meta,
         })
     }
 
@@ -712,23 +723,24 @@ impl SealevelMailboxIndexer {
         let hyperlane_message =
             HyperlaneMessage::read_from(&mut &dispatched_message_account.encoded_message[..])?;
 
-        // let log_meta = self
-        //     .dispatch_message_log_meta(
-        //         U256::from(nonce),
-        //         &valid_message_storage_pda_pubkey,
-        //         &dispatched_message_account.slot,
-        //     )
-        //     .await?;
-
-        let log_meta = LogMeta {
-            address: self.program_id.to_bytes().into(),
-            block_number: dispatched_message_account.slot,
-            // TODO: get these when building out scraper support.
-            // It's inconvenient to get these :|
-            block_hash: H256::zero(),
-            transaction_id: H512::zero(),
-            transaction_index: 0,
-            log_index: U256::zero(),
+        let log_meta = if self.advanced_log_meta {
+            self.dispatch_message_log_meta(
+                U256::from(nonce),
+                &valid_message_storage_pda_pubkey,
+                &dispatched_message_account.slot,
+            )
+            .await?
+        } else {
+            LogMeta {
+                address: self.program_id.to_bytes().into(),
+                block_number: dispatched_message_account.slot,
+                // TODO: get these when building out scraper support.
+                // It's inconvenient to get these :|
+                block_hash: H256::zero(),
+                transaction_id: H512::zero(),
+                transaction_index: 0,
+                log_index: U256::zero(),
+            }
         };
 
         Ok((hyperlane_message.into(), log_meta))
@@ -748,7 +760,7 @@ impl SealevelMailboxIndexer {
         Ok(expected_pubkey)
     }
 
-    async fn _dispatch_message_log_meta(
+    async fn dispatch_message_log_meta(
         &self,
         log_index: U256,
         message_storage_pda_pubkey: &Pubkey,
@@ -805,23 +817,24 @@ impl SealevelMailboxIndexer {
             .into_inner();
         let message_id = delivered_message_account.message_id;
 
-        // let log_meta = self
-        //     .delivered_message_log_meta(
-        //         U256::from(nonce),
-        //         &valid_message_storage_pda_pubkey,
-        //         &delivered_message_account.slot,
-        //     )
-        //     .await?;
-
-        let log_meta = LogMeta {
-            address: self.program_id.to_bytes().into(),
-            block_number: delivered_message_account.slot,
-            // TODO: get these when building out scraper support.
-            // It's inconvenient to get these :|
-            block_hash: H256::zero(),
-            transaction_id: H512::zero(),
-            transaction_index: 0,
-            log_index: U256::zero(),
+        let log_meta = if self.advanced_log_meta {
+            self.delivered_message_log_meta(
+                U256::from(nonce),
+                &valid_message_storage_pda_pubkey,
+                &delivered_message_account.slot,
+            )
+            .await?
+        } else {
+            LogMeta {
+                address: self.program_id.to_bytes().into(),
+                block_number: delivered_message_account.slot,
+                // TODO: get these when building out scraper support.
+                // It's inconvenient to get these :|
+                block_hash: H256::zero(),
+                transaction_id: H512::zero(),
+                transaction_index: 0,
+                log_index: U256::zero(),
+            }
         };
 
         Ok((message_id.into(), log_meta))
@@ -839,7 +852,7 @@ impl SealevelMailboxIndexer {
         Ok(expected_pubkey)
     }
 
-    async fn _delivered_message_log_meta(
+    async fn delivered_message_log_meta(
         &self,
         log_index: U256,
         message_storage_pda_pubkey: &Pubkey,