+++ typescript/infra/config/environments/mainnet3/agent.ts
@@ -22,6 +22,7 @@ import { getDomainId } from '../../registry.js';
 import { environment } from './chains.js';
 import { helloWorld } from './helloworld.js';
 import aaveSenderAddresses from './misc-artifacts/aave-sender-addresses.json';
+import everclearSenderAddresses from './misc-artifacts/everclear-sender-addresses.json';
 import merklyEthAddresses from './misc-artifacts/merkly-eth-addresses.json';
 import merklyNftAddresses from './misc-artifacts/merkly-eth-addresses.json';
 import merklyErc20Addresses from './misc-artifacts/merkly-eth-addresses.json';
@@ -64,7 +65,9 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     bitlayer: true,
     blast: true,
     bob: true,
+    boba: true,
     bsc: true,
+    bsquared: true,
     celo: true,
     cheesechain: true,
     chilizmainnet: true,
@@ -72,6 +75,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     cyber: true,
     degenchain: true,
     dogechain: true,
+    duckchain: true,
     eclipsemainnet: true,
     endurance: true,
     ethereum: true,
@@ -93,6 +97,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     lisk: true,
     lukso: true,
     lumia: true,
+    lumiaprism: true,
     mantapacific: true,
     mantle: true,
     merlin: true,
@@ -124,9 +129,13 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     snaxchain: true,
     solanamainnet: true,
     stride: false,
+    superseed: true,
     superpositionmainnet: true,
+    swell: true,
     taiko: true,
     tangle: true,
+    unichain: true,
+    vana: true,
     viction: true,
     worldchain: true,
     xai: true,
@@ -152,7 +161,9 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     bitlayer: true,
     blast: true,
     bob: true,
+    boba: true,
     bsc: true,
+    bsquared: true,
     celo: true,
     cheesechain: true,
     chilizmainnet: true,
@@ -160,6 +171,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     cyber: true,
     degenchain: true,
     dogechain: true,
+    duckchain: true,
     eclipsemainnet: true,
     endurance: true,
     ethereum: true,
@@ -181,6 +193,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     lisk: true,
     lukso: true,
     lumia: true,
+    lumiaprism: true,
     mantapacific: true,
     mantle: true,
     merlin: true,
@@ -213,9 +226,13 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     snaxchain: true,
     solanamainnet: true,
     stride: true,
+    superseed: true,
     superpositionmainnet: true,
+    swell: true,
     taiko: true,
     tangle: true,
+    unichain: true,
+    vana: true,
     viction: true,
     worldchain: true,
     xai: true,
@@ -241,7 +258,9 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     bitlayer: true,
     blast: true,
     bob: true,
+    boba: true,
     bsc: true,
+    bsquared: true,
     celo: true,
     cheesechain: true,
     chilizmainnet: true,
@@ -249,6 +268,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     cyber: true,
     degenchain: true,
     dogechain: true,
+    duckchain: true,
     // Cannot scrape Sealevel chains
     eclipsemainnet: false,
     endurance: true,
@@ -271,6 +291,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     lisk: true,
     lukso: true,
     lumia: true,
+    lumiaprism: true,
     mantapacific: true,
     mantle: true,
     merlin: true,
@@ -303,9 +324,13 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     // Cannot scrape Sealevel chains
     solanamainnet: false,
     stride: true,
+    superseed: true,
     superpositionmainnet: true,
+    swell: true,
     taiko: true,
     tangle: true,
+    unichain: true,
+    vana: true,
     // Has RPC non-compliance that breaks scraping.
     viction: false,
     worldchain: true,
@@ -395,6 +420,12 @@ const metricAppContextsGetter = (): MetricAppContext[] => {
       name: 'aave',
       matchingList: senderMatchingList(aaveSenderAddresses),
     },
+    {
+      // https://docs.everclear.org/resources/contracts/mainnet
+      // Messages between HubGateway (Everclear hub) <> EverclearSpoke (all other spoke chains)
+      name: 'everclear_gateway',
+      matchingList: senderMatchingList(everclearSenderAddresses),
+    },
   ];
 };
 
@@ -429,7 +460,7 @@ const hyperlane: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: '25a927d-20241114-171323',
+      tag: 'b35c105-20241128-172842',
     },
     gasPaymentEnforcement: gasPaymentEnforcement,
     metricAppContextsGetter,
@@ -438,7 +469,7 @@ const hyperlane: RootAgentConfig = {
   validators: {
     docker: {
       repo,
-      tag: '75d62ae-20241107-060707',
+      tag: 'b35c105-20241128-172842',
     },
     rpcConsensusType: RpcConsensusType.Quorum,
     chains: validatorChainConfig(Contexts.Hyperlane),
@@ -448,7 +479,7 @@ const hyperlane: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: '75d62ae-20241107-060707',
+      tag: 'b35c105-20241128-172842',
     },
     resources: scraperResources,
   },