+++ typescript/infra/config/environments/mainnet3/warp/configGetters/getEthereumInevmUSDTWarpConfig.ts
@@ -2,18 +2,23 @@ import { ethers } from 'ethers';
 
 import {
   ChainMap,
-  RouterConfig,
+  OwnableConfig,
   TokenRouterConfig,
   TokenType,
 } from '@hyperlane-xyz/sdk';
 
-import { tokens } from '../../../../../src/config/warp.js';
+import {
+  RouterConfigWithoutOwner,
+  tokens,
+} from '../../../../../src/config/warp.js';
 
 export const getEthereumInevmUSDTWarpConfig = async (
-  routerConfig: ChainMap<RouterConfig>,
+  routerConfig: ChainMap<RouterConfigWithoutOwner>,
+  abacusWorksEnvOwnerConfig: ChainMap<OwnableConfig>,
 ): Promise<ChainMap<TokenRouterConfig>> => {
   const ethereum: TokenRouterConfig = {
     ...routerConfig.ethereum,
+    ...abacusWorksEnvOwnerConfig.ethereum,
     type: TokenType.collateral,
     token: tokens.ethereum.USDT,
     hook: '0xb87AC8EA4533AE017604E44470F7c1E550AC6F10', // aggregation of IGP and Merkle, arbitrary config not supported for now, TODO: may want to move to zero address in future
@@ -21,6 +26,7 @@ export const getEthereumInevmUSDTWarpConfig = async (
 
   const inevm: TokenRouterConfig = {
     ...routerConfig.inevm,
+    ...abacusWorksEnvOwnerConfig.inevm,
     type: TokenType.synthetic,
     interchainSecurityModule: ethers.constants.AddressZero,
   };