+++ typescript/github-proxy/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @hyperlane-xyz/github-proxy
 
+## 7.2.0
+
+## 7.1.0
+
 ## 7.0.0
 
 ## 6.0.0